msm: camera: isp: fix a cts test failure due to framedrop pattern
The ITS HAL3 test failed due to broken frames and green frames. The
root cause of the issue is due to the framedrop pattern setting in
the isp driver. This change fixes the following issues:

a) Keep tracking the hw framedrop pattern even after the reg update
   ack miss.

b) Remove a race condition between the framedrop pattern setting and
   request frame ioctl handling.

c) Correctly set the framedrop pattern after the request frame.

Change-Id: I862046c718d6e3a28d5f372eb2eb5e371471eb58
Signed-off-by: Jing Zhou <jzhou70@codeaurora.org>