power: qpnp-fg: add support to control slope limiter coefficients
There is a requirement to control the slope limiter coefficients
based on charging status and battery temperature. Add support
for it through "qcom,fg-control-slope-limiter" device tree
property.

Temperature threshold and the slope limiter coefficients can be
specified through the device tree properties if the default
values need to be overridden.

CRs-Fixed: 977693
Change-Id: Ibdb50795352dc572499954a1bda719abdc485a9f
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>