ASoC: wcd-mbhc: Reduce latency during plug correction
Sleeping for a very long time in here leads to slow detections and poor UX.
Instead of sleeping for the same amount of time on each iteration, use a
dynamic delay so that the delay is extended when the number of iterations
goes up. Also don't double check if a stop request has been issued (waste
of time).

Signed-off-by: Sultanxda <sultanxda@gmail.com>