ASoC: wcd9335: Disable haptics while the microphone is open
The noise from haptics is intrusive to the user when the microphone is
open, such as while a video is being recorded. Even when the noise from
haptics isn't a major problem, users typically do not care for haptics
while they're using the microphone.

Disable haptics when a Capture stream is open to fix these issues.

Signed-off-by: Sultanxda <sultanxda@gmail.com>