ARM: dts: 15801/11: Use hardware vsync signal to check panel health
Using ESD_BTA mode (the default) for the panel health status check causes
reboots on devices with the s6e3fa5 panel.

Instead, use the hardware vsync signal to check if the panel is working.
When the vsync interrupt does not fire for 5000 ms (STATUS_CHECK_INTERVAL_MS)
while the screen is on, panel ESD recovery will kick in.

Signed-off-by: Sultanxda <sultanxda@gmail.com>