Revert "msm: mdss: optimize DSI command delay semantics"
None of our panels have a long delay associated with their final power-on
command, so this is not needed.

This is also likely the culprit behind the graphical artifacts observed on
s6e3fa5 panels when powering off the display. If it isn't, then commit
dea3d6c25c08cf8c936475ac690596a83d14366c is.

This reverts commit 78548a7a38ed4c70e8b961352d5b5c70e559bde8.