input: synaptics: Ignore keypad presses when touchscreen is in use
When the user's finger is on the touchscreen, don't process any keypad
button presses, as an unintended button press will cause poor UX.

This ignores home-button presses (fpc1020) and menu/back button
presses (s1302) when the touchscreen (s3320) detects one or more fingers.

Although this is racy, no locks are necessary because synchronization is
not critical for this feature. In order to fix a bug that existed in a
previous implementation (keypad wouldn't work after the screen was turned
on), touch_active is reset when the screen is turned off.

Signed-off-by: Sultanxda <sultanxda@gmail.com>