input: synaptics: s3320: Fix first touch on wake not working sometimes
Release all fingers on suspend so that the first touch when the screen
turns on always works. This fixes the bug where the first touch after the
screen was turned on wouldn't work sometimes.

To reproduce this bug: turn on the screen -> move a finger really quickly
across the screen -> hit the power button to turn off the screen while
moving your finger across it quickly -> take finger off the screen after the
screen turns off -> turn on the screen -> the first touch will be ignored.

Signed-off-by: Sultanxda <sultanxda@gmail.com>